requirements.txt
nvidia-riva-client
ipython
grpc
grpcio
grpcio-tools
numpy
wave
platform
os